feat: add PerpSettingsButton component and integrate it into PerpTickerBar and PerpAccountPanel
